% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorSort.R
\name{sort_vector}
\alias{sort_vector}
\title{Sort a vector}
\usage{
sort_vector(vec)
}
\arguments{
\item{vec}{An integer vector to be sorted.}
}
\description{
This function sorts a given vector using the C++ implementation.
}
